fix(build): Fix Windows MinGW build failures related to udp and midi casts

+++ src/midi.cpp
@@ -1,6 +1,7 @@
 #define MINIMIDIO_IMPLEMENTATION
 #include "../third_party/minimidio/minimidio.h"
 #include <FL/Fl.H>
+#include <stdint.h>
 
 extern void trigger_midi_note(int channel, int note, int velocity);
 extern void release_midi_note(int channel, int note);
@@ -11,21 +12,21 @@ void my_midi_callback(mm_device* dev, const mm_message* msg, void* userdata) {
         int vel = msg->data[1];
         int ch = msg->channel;
         Fl::awake([](void* data) {
-            long val = (long)data;
+            intptr_t val = (intptr_t)data;
             int c = (val >> 16) & 0xFF;
             int n = (val >> 8) & 0xFF;
             int v = val & 0xFF;
             trigger_midi_note(c, n, v);
-        }, (void*)((long)((ch << 16) | (note << 8) | vel)));
+        }, (void*)((intptr_t)((ch << 16) | (note << 8) | vel)));
     } else if (msg->type == MM_NOTE_OFF || (msg->type == MM_NOTE_ON && msg->data[1] == 0)) {
         int note = msg->data[0];
         int ch = msg->channel;
         Fl::awake([](void* data) {
-            long val = (long)data;
+            intptr_t val = (intptr_t)data;
             int c = (val >> 16) & 0xFF;
             int n = (val >> 8) & 0xFF;
             release_midi_note(c, n);
-        }, (void*)((long)((ch << 16) | (note << 8))));
+        }, (void*)((intptr_t)((ch << 16) | (note << 8))));
     }
 }
